+++ src/services/api/auth/logout.js
@@ -1,4 +1,4 @@
-import api from '../../lib/api';
+import api from '../../../lib/api';
 
 const logoutAPI = () => {
   return api.post(
